third_party/WebKit/LayoutTests/external/wpt/css/css-animations/CSSPseudoElement-getAnimations.tentative-expected.txt

This is a testharness.js-based test.
FAIL getAnimations returns CSSAnimation objects assert_equals: CSS animation name matches expected (string) "anim1" but got (undefined) undefined
FAIL getAnimations returns CSS transitions/animations, and script-generated animations in the expected order assert_equals: 1st animation is the 1st transition sorted by name expected (string) "height" but got (undefined) undefined
Harness: the test ran to completion.
